+++ bundlers/browser/TsBrowserify.ts
@@ -139,13 +139,10 @@ class TsBrowserify extends EventEmitter.EventEmitter {
     constructor(files: any, opts: BrowserifyOptions);
     constructor(files: any, options?: BrowserifyOptions) {
         super();
-        if (options == null) {
-            var fileOpts = <BrowserifyOptions>files;
-            if (fileOpts.insertModuleGlobals != null || fileOpts.moduleDeps != null || fileOpts.basedir != null) {
-                options = fileOpts;
-            }
-        }
+        options = (options != null ? options : <BrowserifyOptions>files);
         if (options == null) throw new Error("'options' is required");
+        if (options.browserPack == null) throw new Error("'options.browserPack' is required");
+        if (options.depsSort == null) throw new Error("'options.depsSort' is required");
         if (options.insertModuleGlobals == null) throw new Error("'options.insertModuleGlobals' is required");
         if (options.moduleDeps == null) throw new Error("'options.moduleDeps' is required");